Refactor to prepare for running our demo on other platforms
